✨Finished Appliance methods

+++ src/main/java/com/warm/service/ApplianceServiceImpl.java
@@ -26,8 +26,18 @@ public class ApplianceServiceImpl implements ApplianceService {
     private UserRepository userRepository;
 
     @Override
+    public Appliance findByIdAndUserId(Long id, Long userId) {
+        if (!userRepository.existsById(userId)) {
+            throw new ServiceException(Error.USER_NOT_FOUND);
+        } else {
+            return applianceRepository.findById(id).orElseThrow(() -> new ServiceException(Error.APPLIANCE_NOT_FOUND));
+        }
+    }
+
+    @Override
     public Appliance findById(Long id) {
-        return applianceRepository.findById(id).orElseThrow(() -> new ServiceException(Error.APPLIANCE_NOT_FOUND));
+        return applianceRepository.findById(id).orElseThrow(
+                () -> new ServiceException(Error.APPLIANCE_NOT_FOUND));
     }
 
     @Override
@@ -40,36 +50,48 @@ public class ApplianceServiceImpl implements ApplianceService {
     }
 
     @Override
-    public Appliance update(Long id, SaveApplianceRequest applianceReq) {
-        return applianceRepository.findById(id).map(appliance -> {
-            appliance.setName(applianceReq.getName());
-            appliance.setBrand(applianceReq.getBrand());
-            appliance.setModel(applianceReq.getModel());
-            appliance.setType(applianceReq.getType());
-            return applianceRepository.save(appliance);
-        }).orElseThrow(() -> new ServiceException(Error.APPLIANCE_NOT_FOUND));
+    public Appliance update(Long id, SaveApplianceRequest request) {
+        if (!userRepository.existsById(request.getUserId())){
+            throw new ServiceException(Error.USER_NOT_FOUND);
+        } else {
+            return applianceRepository.findById(id).map(appliance -> {
+                appliance.setName(request.getName());
+                appliance.setBrand(request.getBrand());
+                appliance.setModel(request.getModel());
+                appliance.setType(request.getType());
+                return applianceRepository.save(appliance);
+            }).orElseThrow(() -> new ServiceException(Error.APPLIANCE_NOT_FOUND));
+        }
     }
 
     @Override
-    public ResponseEntity<?> delete(Long id) {
-        return applianceRepository.findById(id).map(appliance -> {
-            applianceRepository.delete(appliance);
-            return ResponseEntity.ok().build();
-        }).orElseThrow(() -> new ServiceException(Error.APPLIANCE_NOT_FOUND));
+    public ResponseEntity<?> delete(Long id, Long userId) {
+        if (!userRepository.existsById(userId)){
+            throw new ServiceException(Error.USER_NOT_FOUND);
+        } else{
+            return applianceRepository.findById(id).map(appliance -> {
+                applianceRepository.delete(appliance);
+                return ResponseEntity.ok().build();
+            }).orElseThrow(() -> new ServiceException(Error.APPLIANCE_NOT_FOUND));
+        }
     }
 
     @Override
     public Appliance create(SaveApplianceRequest request) {
-        Appliance appliance = fromReq(request);
-        appliance.setUser(userService.findById(request.getUserId()));
-        return applianceRepository.save(appliance);
+        if (!userRepository.existsById(request.getUserId())){
+            throw new ServiceException(Error.USER_NOT_FOUND);
+        } else {
+            Appliance appliance = fromReq(request);
+            appliance.setUser(userService.findById(request.getUserId()));
+            return applianceRepository.save(appliance);
+        }
     }
 
     //mapper
     private Appliance fromReq(SaveApplianceRequest request) {
         Appliance appliance;
         if (request.getId() != null) {
-            appliance = findById(request.getId());
+            appliance = findByIdAndUserId(request.getId(), request.getUserId());
         } else {
             appliance = Appliance.builder().build();
         }
@@ -81,6 +103,7 @@ public class ApplianceServiceImpl implements ApplianceService {
         return appliance;
     }
 
+    /*
     public ApplianceResource convertToResource(Appliance appliance) {
         ApplianceResource applianceResource = new ApplianceResource();
 
@@ -90,5 +113,5 @@ public class ApplianceServiceImpl implements ApplianceService {
         applianceResource.setType(appliance.getType());
 
         return applianceResource;
-    }
+    }*/
 }
